choose_disk: Unificate input parameters by using vars
The actual code was expecting a set of variables (facts, disks,legacy_disks).
This was pretty painful as for every new needed variable, it was
necessary to add a new entry point. That made also the variable passing
less intuitive as switching from native to legacy requires to edit the
code in two places.

This patch is about using {{vars}} directly in the ansible module.
The main benefit is reducing the list of input variables.
The drawback is that we should double check the type & the variable
presence inside the module itself.

The patch is adding get_vars() function as an helper to reach any
variable defined in vars. If the lookup variable doesn't exist, None is
returned.

Since we use {{vars}} as a global input, the variable typing is done in
our code allowing detecting if "devices" is a list or a dict.

The goodness of that is being able to detect if we have the legacy or
the native syntax.

This patch is also about having a single "devices" input variable which makes the
transition easier between the legacy & native syntax. Another benefit
is having a simplier documentation & code.

The raw_journal_devices is following the same principle and will be
passed through the 'vars'.

Note that ansible variables will have the same name in this python
module like in :

    devices = get_var(module, "devices")