choose_disk: Improve unit size parsing
The actual code was unable to parse input like 'gt(30GB)', that's not
user friendly at all.

This patch does improve the parsing by not trying to read the unit twice
from the input.

The size information is just removed from the original string.

Adding a unit test to check that case.